+++ packages/web3-eth/src/rpc_methods.ts
@@ -12,6 +12,7 @@ import {
 } from 'web3-utils';
 import { validator } from 'web3-validator';
 import { validateTransactionCall, validateTransactionWithSender } from './validation';
+import { Web3EthExecutionAPI } from './web3_eth_execution_api';
 
 export async function getProtocolVersion(requestManager: Web3RequestManager) {
 	return requestManager.send({
@@ -494,21 +495,23 @@ export async function getFeeHistory(
 	});
 }
 
-export async function getPendingTransactions(requestManager: Web3RequestManager) {
+export async function getPendingTransactions(
+	requestManager: Web3RequestManager<Web3EthExecutionAPI>,
+) {
 	return requestManager.send({
 		method: 'eth_pendingTransactions',
 		params: [],
 	});
 }
 
-export async function requestAccounts(requestManager: Web3RequestManager) {
+export async function requestAccounts(requestManager: Web3RequestManager<Web3EthExecutionAPI>) {
 	return requestManager.send({
 		method: 'eth_requestAccounts',
 		params: [],
 	});
 }
 
-export async function getChainId(requestManager: Web3RequestManager) {
+export async function getChainId(requestManager: Web3RequestManager<Web3EthExecutionAPI>) {
 	return requestManager.send({
 		method: 'eth_chainId',
 		params: [],
@@ -516,7 +519,7 @@ export async function getChainId(requestManager: Web3RequestManager) {
 }
 
 export async function getProof(
-	requestManager: Web3RequestManager,
+	requestManager: Web3RequestManager<Web3EthExecutionAPI>,
 	address: Address,
 	storageKey: HexString32Bytes,
 	blockNumber: BlockNumberOrTag,
